requirements.txt
fastapi
uvicorn
streamlit
yfinance
requests
beautifulsoup4
pandas
faiss-cpu
langchain
whisper
coqui-tts
sentence-transformers
transformers
torch
sounddevice
scipy
pyttsx3
gtts
pygame


